chore(lint): take the golangci baseline to zero
The lint job runs with continue-on-error and reported pass at 131
findings, so every PR's annotation panel showed ten of main's issues and
none of its own.

Fixed: gofmt/gofumpt/goimports across 48 files; reflect.Ptr → reflect.Pointer
(govet inline); three unconvert; two Deprecated notices in their own
paragraph; slices.Concat for the append-to-other-slice in spec.go; an
ineffectual assignment in a chainstate test; a dead err == nil fallback
after LavaFormatTrace, which always returns an error; two multi-value
composite-literal returns that gofmt and gofumpt indent differently, given
names so both agree. Test doubles that return nil, nil and the typed-nil
assertion carry a nolint with the reason.

Excluded in .golangci.yml, each with the reason: SA1019 for
jhump/protoreflect (a protobuf v1 → v2 reflection migration) and
x/net/http2/h2c (superseded by http.Server.Protocols; the cleartext
HTTP/2 path moves in its own change); ST1005 (error strings are
wire-visible) and ST1012 (23 exported error variable renames). The
golang/protobuf exclusion matches staticcheck's unquoted message form.

+++ protocol/lavasession/endpoint_relay_probe_test.go
@@ -55,7 +55,7 @@ func TestRelayProbe_GatesReEnableUntilConfirmed(t *testing.T) {
 	require.True(t, e.ConfirmRelayRecovery())
 	require.True(t, e.Enabled)
 	e.mu.RLock()
-	require.Equal(t, uint64(MaxConsecutiveConnectionAttempts-probeReenableTrialBudget), e.ConnectionRefusals,
+	require.Equal(t, MaxConsecutiveConnectionAttempts-probeReenableTrialBudget, e.ConnectionRefusals,
 		"a replay-confirmed re-enable still carries only the trial budget — one replayed request is not real traffic")
 	require.True(t, e.probeReenabled, "a replay-confirmed re-enable is still probe-granted until real traffic validates it")
 	e.mu.RUnlock()
@@ -271,7 +271,7 @@ func TestRelayProbe_EvidenceDroppedAfterMaxFailedReplays(t *testing.T) {
 	}
 	require.True(t, e.IsEnabled(), "after the evidence is dropped, poll hysteresis alone must re-enable the endpoint")
 	e.mu.RLock()
-	require.Equal(t, uint64(MaxConsecutiveConnectionAttempts-probeReenableTrialBudget), e.ConnectionRefusals,
+	require.Equal(t, MaxConsecutiveConnectionAttempts-probeReenableTrialBudget, e.ConnectionRefusals,
 		"the fallback re-enable still carries only the trial budget")
 	e.mu.RUnlock()
 }
